friendlyTask/mysite/requirements.txt
cfscrape==2.1.1
cloudscraper==1.2.64
colorthief==0.2.1
Django==2.2.12
djangorestframework==3.14.0
Pillow==9.2.0
requests==2.22.0
selenium==4.5.0
undetected_chromedriver==3.1.6
